Extract admin session helpers to concern
This makes the code related to checking and ending expired admin
sessions available more widely, specifically so that it can be used by
the session refresh code to avoid extending expired sessions (coming
next).
