+++ app/controllers/admin/base_admin_controller.rb
@@ -1,6 +1,6 @@
 module Admin
   class BaseAdminController < ApplicationController
-    ADMIN_TIMEOUT_LENGTH_IN_MINUTES = 30
+    include AdminSessionTimeout
 
     layout "admin"
 
@@ -8,27 +8,6 @@ module Admin
     helper_method :admin_signed_in?, :admin_timeout_in_minutes, :service_operator_signed_in?
 
     private
-
-    def admin_signed_in?
-      session.key?(:user_id)
-    end
-
-    def admin_timeout_in_minutes
-      ADMIN_TIMEOUT_LENGTH_IN_MINUTES
-    end
-
-    def admin_session_timed_out?
-      admin_signed_in? && session[:last_seen_at] < admin_timeout_in_minutes.minutes.ago
-    end
-
-    def end_expired_admin_sessions
-      if admin_session_timed_out?
-        session.delete(:user_id)
-        session.delete(:organisation_id)
-        session.delete(:role_codes)
-        flash[:notice] = "Your session has timed out due to inactivity, please sign-in again"
-      end
-    end
 
     def ensure_authenticated_user
       unless admin_signed_in?
